Update dependencies and workflow in Django.yml
Switched the pip install command to use the general 'requirements.txt' instead of the 'requirements_production.txt'. Also, modified the Flake8 Lint workflow to point at the 'edc-data-manager' directory instead of 'potlako'. This update aligns the workflow with the correct dependency file and ensures linting is run against the primary project directory.