Updated Flake8 Lint path in django.yml
Adjusted the path for the Flake8 Lint run command in the Github workflow file 'django.yml' from the specific 'edc-data-manager' directory to cover the entire workspace, ensuring linting is performed across all code included in the repository, not limited to just the 'edc-data-manager' directory. Now, it is working in a more comprehensive manner.